refactor: separate mobile and desktop selection gesture detector
