+++ lib/src/editor/editor_component/service/selection/mobile_selection_service.dart
@@ -1,7 +1,7 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_editor/src/flutter/overlay.dart';
 import 'package:appflowy_editor/src/render/selection/mobile_selection_widget.dart';
-import 'package:appflowy_editor/src/service/selection/mobile_selection_gesture.dart';
+import 'package:appflowy_editor/src/service/selection/mobile_selection_gesture_detector.dart';
 import 'package:flutter/material.dart' hide Overlay, OverlayEntry;
 
 import 'package:appflowy_editor/src/render/selection/cursor_widget.dart';
