+++ lib/src/service/selection/mobile_selection_gesture_detector.dart
